// src/core/database/interfaces/dbConfig.interface.ts
export interface IDatabaseConfigAttributes {
  username?: string;
  password?: string;
  database?: string;
  host?: string;
  port?: number | string;
  dialect?: string;
  urlDatabase?: string;
}

export interface IDatabaseConfig {
  development: IDatabaseConfigAttributes;
  test: IDatabaseConfigAttributes;
  staging: IDatabaseConfigAttributes;
  production: IDatabaseConfigAttributes;
}